+++ src/main/java/br/com/Project/Pack2/Subtrair.java
@@ -1,7 +1,7 @@
 package br.com.Project.Pack2;
 
 public class Subtrair {
-    public int subitrai (int a, int b){
+    public double subitrai (double a, double b){
         return a - b;
     }
 }
